pyproject.toml
[project]
name = "ate-features"
version = "0.1.0"
requires-python = ">=3.11"
description = "Agent Teams Eval: feature implementation in LangGraph"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.15",
]

[project.scripts]
ate-features = "ate_features.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-timeout>=2.4",
    "ruff>=0.15",
    "mypy>=1.19",
    "types-PyYAML>=6.0",
]
scoring = [
    "rich>=14.0",
]
analysis = [
    "pandas>=2.2",
    "matplotlib>=3.10",
    "scipy>=1.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ate_features"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
